m_main.c is now a glibc-free zone!
It's becoming clearer how startup should be structured.  Basically the
address space manager should be started as early as possible -- much
earlier than it is at present, before any poking around at command
lines, environment variables or whatever.  That's because so much of
the startup stuff depends on dynamic memory allocation, which depends
on aspacemgr.