Made ExeContext gathering more efficient.  Previously we were getting as many code addresses as possible, and then filling the rest of the array (up to VG_(clo_backtrace_size)) with zeroes.  These zero entries were then included in the hashing, and comparisons, and space was allocated for them in saved ExeContexts.
By not putting in the zeroes, not doing any hashing/comparisons of them, and
not storing them, we see speed-ups for perf/heap of 5% with
--num-callers=12 and 7% for --num-callers=50, and about 1.5% for tinycc.
The amount of memory for ExeContexts also drops, saving around 500KB for
tinycc with --num-callers=50.

I also changed the allocation of each ExeContext to use "sizeof(struct
_ExeContext)" instead of "sizeof(struct _ExeContext *)".  It worked ok
up until now because the struct only contained a single pointer in the
non-variable-sized part, but it was an accident waiting to happen when
struct _ExeContext changed (and indeed did happen to me when I added
'n_ips' to the struct).