Initial implementation of VTS pruning.  This removes VTS entries for threads that have exited, thereby limiting the size of the VTSes to O(# currently live threads) rather than O(# threads that have ever lived).  This drastically improves behaviour on programs that create many threads over their lifetime, but do not have many live threads at any given time.  Downside is some (hopefully small) risk of missing races.  Controlled by new flag --vts-pruning=never|auto|always [auto].
Since it requires visiting each VtsID in the system, this gives for
free an opportunity for cross-checking the refcounts of the VtsIDs.

TODO: add documentation about this
TODO: more principled way of deciding when a thread can be pruned
TODO: better heuristics for deciding when to do pruning

Also rearrange data decls in libhb_core.c; move the decls up to the
top of the file, away from their associated functions.